requirements.txt
matplotlib==3.4.3
numpy==1.21.2
open3d==0.14.1
pandas==1.3.4
Pillow==9.2.0
plotly==5.6.0
scikit_image==0.18.3
scikit_learn==1.1.2
scipy==1.7.3
skimage==0.0
torch==1.1.0
torchvision==0.8.1
